+++ js/charts/boxplot.js
@@ -136,7 +136,7 @@ function boxplot() {
         }
     }
 
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
@@ -144,9 +144,11 @@ function boxplot() {
         }
     }
 
-    var clearFilter = function () {
+    var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
         }
     }
 
@@ -498,10 +500,10 @@ function boxplot() {
             UTIL.setAxisColor(_local_svg, "", "", true, true, true, true);
 
             d3.select(div).select('.filterData')
-                .on('click', applyFilter(chart));
+                .on('click', applyFilter(div));
 
             d3.select(div).select('.removeFilter')
-                .on('click', clearFilter());
+                .on('click', clearFilter(div));
 
             var lasso = d3
                 .lasso()
