+++ js/charts/bullet.js
@@ -135,7 +135,7 @@ function bullet() {
         }
     }
 
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
@@ -143,9 +143,11 @@ function bullet() {
         }
     }
 
-    var clearFilter = function () {
+   var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
         }
     }
 
@@ -423,10 +425,10 @@ function bullet() {
             formatUsingCss(this);
 
             d3.select(div).select('.filterData')
-                .on('click', applyFilter(chart));
+                .on('click', applyFilter(div));
 
             d3.select(div).select('.removeFilter')
-                .on('click', clearFilter());
+                .on('click', clearFilter(div));
 
             var lasso = d3.lasso()
                 .hoverSelect(true)
