+++ js/charts/clusteredverticalbar.js
@@ -1,9 +1,10 @@
-var COMMON = require('../extras/common.js')(),
-    UTIL = require('../extras/util.js')(),
-    LEGEND = require('../extras/legend_barcharts.js')();
+//var d3 = require('d3');
+var COMMON = require('../extras/common.js')();
+var UTIL = require('../extras/util.js')();
+var LEGEND = require('../extras/legend_barcharts.js')();
+//var LASSO = require('d3-lasso'); // creates a new lasso
 
 function clusteredverticalbar() {
-
     var _NAME = 'clusteredverticalbar';
 
     var _config,
@@ -19,7 +20,7 @@ function clusteredverticalbar() {
         _showYaxisLabel,
         _xAxisColor,
         _yAxisColor,
-        _showGrid ,
+        _showGrid,
         _stacked,
         _displayName,
         _legendData,
@@ -31,14 +32,15 @@ function clusteredverticalbar() {
         _textColor = [],
         _displayColor = [],
         _borderColor = [],
-        _fontSize = [];
+        _fontSize = [],
+        _print;
 
     var _local_svg, _Local_data, _originalData, _localLabelStack = [], legendBreakCount = 1;
     var _localXAxis,
         _localYAxis,
         _localXGrid,
         _localYGrid;
-    var parentWidth, parentHeight, plotWidth, plotHeight, container;
+    var parentWidth, parentHeight, plotWidth, plotHeight, container, tooltip;
 
     var x0, x1, _xDimensionGrid = d3.scaleLinear(), y;
     var margin = {
@@ -68,6 +70,7 @@ function clusteredverticalbar() {
         this.showYaxis(config.showYaxis);
         this.showXaxisLabel(config.showXaxisLabel);
         this.showYaxisLabel(config.showYaxisLabel);
+        this.showGrid(config.showGrid);
         this.xAxisColor(config.xAxisColor);
         this.yAxisColor(config.yAxisColor);
         this.displayName(config.displayName);
@@ -88,7 +91,7 @@ function clusteredverticalbar() {
     var _buildTooltipData = function (datum, chart) {
         var output = "";
 
-        output += "<table><tr>"
+        output += "<table data-toggle='tooltip'><tr>"
             + "<th>" + chart.dimension() + ": </th>"
             + "<td>" + datum[chart.dimension()] + "</td>"
             + "</tr><tr>"
@@ -165,23 +168,25 @@ function clusteredverticalbar() {
         }
     }
 
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
             }
         }
     }
-    var clearFilter = function () {
+    var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
+           
         }
     }
     var _handleMouseOverFn = function (tooltip, container) {
         var me = this;
         return function (d, i) {
             d3.select(this).style('cursor', 'pointer')
-                .style('cursor', 'pointer')
                 .style('fill', COMMON.HIGHLIGHTER);
             var border = UTIL.getDisplayColor(_measure.indexOf(d.measure), _displayColor)
             if (tooltip) {
@@ -263,7 +268,10 @@ function clusteredverticalbar() {
             var _filter = UTIL.createFilterElement()
             $(div).append(_filter);
 
-            $(document).on('click', '_local_svg', function (e) {
+            // var _alert = UTIL.createAlertElement()
+            // $(div).append(_alert);
+
+            $(document).on('click', _local_svg, function (e) {
                 if ($("#myonoffswitch").prop('checked') == false) {
                     var element = e.target
                     if (element.tagName == "_local_svg") {
@@ -273,6 +281,19 @@ function clusteredverticalbar() {
                         $('#Modal_' + $(div).attr('id')).modal('toggle');
                     }
                 }
+            })
+
+            _local_svg.on('click', function (e, i) {
+                if ($(this).parent().find('.alert').prop('checked') == false) {
+                    var element = e.target
+                    if (element.tagName == "_local_svg") {
+                        $('#Modal_' + $(div).attr('id') + ' .measure').val('')
+                        $('#Modal_' + $(div).attr('id') + ' .threshold').val('')
+                        $('#Modal_' + $(div).attr('id') + ' .measure').attr('disabled', false)
+                        $('#Modal_' + $(div).attr('id')).modal('toggle');
+                    }
+                }
+
             })
 
             $(document).on('click', '#Modal_' + $(div).attr('id') + ' .ThresholdSubmit', function (e) {
@@ -304,14 +325,14 @@ function clusteredverticalbar() {
                 legendBreakCount = result.legendBreakCount;
 
                 switch (_legendPosition) {
-                    case 'top':
+                    case 'Top':
                         plotHeight = parentHeight - legendHeight - axisLabelSpace;
                         break;
-                    case 'bottom':
+                    case 'Bottom':
                         plotHeight = parentHeight - legendHeight - axisLabelSpace * 2;
                         break;
-                    case 'right':
-                    case 'left':
+                    case 'Right':
+                    case 'Left':
                         plotWidth = parentWidth - legendWidth;
                         break;
                 }
@@ -367,7 +388,6 @@ function clusteredverticalbar() {
 
         y = d3.scaleLinear()
             .rangeRound([plotHeight, 0]);
-
 
         var plot = container.append('g')
             .attr('class', 'clusteredverticalbar-plot')
@@ -463,7 +483,6 @@ function clusteredverticalbar() {
 
         var rect = drawViz(clusteredverticalbar);
 
-
         var text = clusteredverticalbar.append('text')
             .text(function (d, i) {
                 return UTIL.getFormattedValue(d[d.measure], UTIL.getValueNumberFormat(i, _numberFormat));
@@ -506,64 +525,78 @@ function clusteredverticalbar() {
         var xAxisGroup,
             yAxisGroup;
 
+        var isRotate = false;
 
-        _localXAxis = d3.axisBottom(x0)
-            .tickSize(0)
-            .tickPadding(10);
+        if (_showXaxis) {
+            _localXAxis = d3.axisBottom(x0)
+                .tickSize(0)
+                .tickFormat(function (d) {
+                    if (isRotate == false) {
+                        isRotate = UTIL.getTickRotate(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                    }
+                    return UTIL.getTruncatedTick(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                })
+                .tickPadding(10);
 
-        xAxisGroup = plot.append('g')
-            .attr('class', 'x axis')
-            .attr('visibility', function () {
-                return _showXaxis;
-            })
-            .attr('transform', 'translate(0, ' + plotHeight + ')')
-            .call(_localXAxis);
+            xAxisGroup = plot.append('g')
+                .attr('class', 'x axis')
+                .attr('visibility', function () {
+                    return 'visible';
+                })
+                .attr('transform', 'translate(0, ' + plotHeight + ')')
+                .call(_localXAxis);
 
-        xAxisGroup.append('g')
-            .attr('class', 'label')
-            .attr('transform', function () {
-                return 'translate(' + (plotWidth) + ', ' + (COMMON.AXIS_THICKNESS / 1.5) + ')';
-            })
-            .append('text')
-            .style('text-anchor', 'end')
-            .style('font-weight', 'bold')
-            .style('fill', _xAxisColor)
-            .attr('visibility', function () {
-                return _showXaxisLabel;
-            })
-            .text(_displayName);
+            xAxisGroup.append('g')
+                .attr('class', 'label')
+                .attr('transform', function () {
+                    return 'translate(' + (plotWidth) + ', ' + (COMMON.AXIS_THICKNESS / 1.5) + ')';
+                })
+                .append('text')
+                .style('text-anchor', 'end')
+                .style('font-weight', 'bold')
+                .style('fill', _xAxisColor)
+                .text(_displayName);
 
-        _setAxisColor(xAxisGroup, _xAxisColor);
+            if (isRotate) {
+                _local_svg.selectAll('.x .tick text')
+                    .attr("transform", "rotate(-15)");
+            }
 
-        _localYAxis = d3.axisLeft(y)
-            .tickSize(0)
-            .tickPadding(8)
-            .tickFormat(function (d) {
-                return UTIL.shortScale(2)(d);
-            });
+            _setAxisColor(xAxisGroup, _xAxisColor);
 
-        yAxisGroup = plot.append('g')
-            .attr('class', 'y axis')
-            .attr('visibility', _showYaxis)
-            .call(_localYAxis);
+        }
 
-        yAxisGroup.append('g')
-            .attr('class', 'label')
-            .attr('transform', function () {
-                return 'translate(' + (-COMMON.AXIS_THICKNESS / 1.15) + ', ' + '0)';
-            })
-            .append('text')
-            .attr('transform', 'rotate(-90)')
-            .style('text-anchor', 'end')
-            .style('font-weight', 'bold')
-            .style('fill', _yAxisColor)
-            .attr('visibility', _showYaxisLabel)
-            .text(function () {
-                return _displayNameForMeasure.map(function (p) { return p; }).join(', ');
-            });
+        if (_showYaxis) {
+            _localYAxis = d3.axisLeft(y)
+                .tickSize(0)
+                .tickPadding(8)
+                .tickFormat(function (d) {
+                    return UTIL.shortScale(2)(d);
+                });
 
-        _setAxisColor(yAxisGroup, _yAxisColor);
+            yAxisGroup = plot.append('g')
+                .attr('class', 'y axis')
+                .attr('visibility', function () {
+                    return 'visible';
+                })
+                .call(_localYAxis);
 
+            yAxisGroup.append('g')
+                .attr('class', 'label')
+                .attr('transform', function () {
+                    return 'translate(' + (-COMMON.AXIS_THICKNESS / 1.15) + ', ' + '0)';
+                })
+                .append('text')
+                .attr('transform', 'rotate(-90)')
+                .style('text-anchor', 'end')
+                .style('font-weight', 'bold')
+                .style('fill', _yAxisColor)
+                .text(function () {
+                    return _displayNameForMeasure.map(function (p) { return p; }).join(', ');
+                });
+
+            _setAxisColor(yAxisGroup, _yAxisColor);
+        }
 
         _local_svg.select('g.sort').remove();
         UTIL.sortingView(container, parentHeight, parentWidth + margin.left, legendBreakCount, axisLabelSpace, offsetX);
@@ -582,16 +615,17 @@ function clusteredverticalbar() {
                         $(me).parent().find('.sort_selection,.arrow-down').css('visibility', 'hidden');
                         _local_svg.select('.plot').remove()
                         drawPlot.call(me, _Local_data);
-                        break; arrow - down
+
+                        break;
                     }
                 }
             });
 
         d3.select(div).select('.filterData')
-            .on('click', applyFilter(chart));
+            .on('click', applyFilter(div));
 
         d3.select(div).select('.removeFilter')
-            .on('click', clearFilter());
+            .on('click', clearFilter(div));
 
         _local_svg.select('g.lasso').remove();
 
@@ -612,77 +646,104 @@ function clusteredverticalbar() {
 
     var drawViz = function (element) {
         var me = this;
-
-        element.append('rect')
-            .attr("x", function (d) {
-                return 0;
-            })
-            .attr("y", function (d) {
-                return y(d[d.measure]);
-            })
-            .attr('class', 'bar')
-            .attr("width", x1.bandwidth())
-            .attr("height", function (d) {
-                return 0;
-            })
-            .style('fill', function (d, i) {
-                return UTIL.getDisplayColor(_measure.indexOf(d.measure), _displayColor);
-            })
-            .style('stroke', function (d, i) {
-                return UTIL.getBorderColor(_measure.indexOf(d.measure), _borderColor);
-            })
-            .style('stroke-width', 2)
-            .on('mouseover', _handleMouseOverFn.call(chart, tooltip, _local_svg))
-            .on('mousemove', _handleMouseMoveFn.call(chart, tooltip, _local_svg))
-            .on('mouseout', _handleMouseOutFn.call(chart, tooltip, _local_svg))
-            .on('click', function (d) {
-                if ($("#myonoffswitch").prop('checked') == false) {
-                    $('#Modal_' + $(div).attr('id') + ' .measure').val(d.measure);
-                    $('#Modal_' + $(div).attr('id') + ' .threshold').val('');
-                    $('#Modal_' + $(div).attr('id') + ' .measure').attr('disabled', true);;
-                    $('#Modal_' + $(div).attr('id')).modal('toggle');
-                }
-                else {
-                    var confirm = d3.select(div).select('.confirm')
-                        .style('visibility', 'visible');
-                    var _filter = _Local_data.filter(function (d1) {
-                        return d[_dimension[0]] === d1[_dimension[0]]
-                    })
-                    var rect = d3.select(this);
-                    if (rect.classed('selected')) {
-                        rect.classed('selected', false);
-                        filterData.map(function (val, i) {
-                            if (val[_dimension[0]] == d[_dimension[0]]) {
-                                filterData.splice(i, 1)
-                            }
+        if (!_print) {
+            element.append('rect')
+                .attr("x", function (d) {
+                    return 0;
+                })
+                .attr("y", function (d) {
+                    return y(d[d.measure]);
+                })
+                .attr('class', 'bar')
+                .attr("width", x1.bandwidth())
+                .attr("height", function (d) {
+                    return 0;
+                })
+                .style('fill', function (d, i) {
+                    return UTIL.getDisplayColor(_measure.indexOf(d.measure), _displayColor);
+                })
+                .style('stroke', function (d, i) {
+                    return UTIL.getBorderColor(_measure.indexOf(d.measure), _borderColor);
+                })
+                .style('stroke-width', 2)
+                .on('mouseover', _handleMouseOverFn.call(chart, tooltip, _local_svg))
+                .on('mousemove', _handleMouseMoveFn.call(chart, tooltip, _local_svg))
+                .on('mouseout', _handleMouseOutFn.call(chart, tooltip, _local_svg))
+                .on('click', function (d) {
+                    if ($("#myonoffswitch").prop('checked') == false) {
+                        $('#Modal_' + $(div).attr('id') + ' .measure').val(d.measure);
+                        $('#Modal_' + $(div).attr('id') + ' .threshold').val('');
+                        $('#Modal_' + $(div).attr('id') + ' .measure').attr('disabled', true);;
+                        $('#Modal_' + $(div).attr('id')).modal('toggle');
+                    }
+                    else {
+                        var confirm = d3.select(div).select('.confirm')
+                            .style('visibility', 'visible');
+                        var _filter = _Local_data.filter(function (d1) {
+                            return d[_dimension[0]] === d1[_dimension[0]]
                         })
-                    } else {
-                        rect.classed('selected', true);
-                        var isExist = filterData.filter(function (val) {
-                            if (val[_dimension[0]] == d[_dimension[0]]) {
-                                return val
+                        var rect = d3.select(this);
+                        if (rect.classed('selected')) {
+                            rect.classed('selected', false);
+                            filterData.map(function (val, i) {
+                                if (val[_dimension[0]] == d[_dimension[0]]) {
+                                    filterData.splice(i, 1)
+                                }
+                            })
+                        } else {
+                            rect.classed('selected', true);
+                            var isExist = filterData.filter(function (val) {
+                                if (val[_dimension[0]] == d[_dimension[0]]) {
+                                    return val
+                                }
+                            })
+                            if (isExist.length == 0) {
+                                filterData.push(_filter[0]);
                             }
-                        })
-                        if (isExist.length == 0) {
-                            filterData.push(_filter[0]);
                         }
                     }
-                }
-            })
-            .transition()
-            .duration(COMMON.DURATION)
-            .attr("height", function (d, i) {
-                return plotHeight - y(d[d.measure]);
-            })
-            .attr("y", function (d, i) {
-                return y(d[d.measure]);
-            })
-            .attr("width", x1.bandwidth())
-            .attr("x", function (d, i) {
-                return x1(d.measure);;
-            })
+                })
+                .transition()
+                .duration(COMMON.DURATION)
+                .attr("height", function (d, i) {
+                    return plotHeight - y(d[d.measure]);
+                })
+                .attr("y", function (d, i) {
+                    return y(d[d.measure]);
+                })
+                .attr("width", x1.bandwidth())
+                .attr("x", function (d, i) {
+                    return x1(d.measure);;
+                })
+        }
+        else {
+            element.append('rect')
+                .attr('class', 'bar')
+                .style('fill', function (d, i) {
+                    return UTIL.getDisplayColor(_measure.indexOf(d.measure), _displayColor);
+                })
+                .style('stroke', function (d, i) {
+                    return UTIL.getBorderColor(_measure.indexOf(d.measure), _borderColor);
+                })
+                .style('stroke-width', 2)
+                .attr("height", function (d, i) {
+                    return plotHeight - y(d[d.measure]);
+                })
+                .attr("y", function (d, i) {
+                    return y(d[d.measure]);
+                })
+                .attr("width", x1.bandwidth())
+                .attr("x", function (d, i) {
+                    return x1(d.measure);;
+                })
+        }
+
     }
     chart._legendInteraction = function (event, data, plot) {
+        if (_print) {
+            // No interaction during print enabled
+            return;
+        }
         switch (event) {
             case 'mouseover':
                 _legendMouseOver(data, plot);
@@ -731,6 +792,10 @@ function clusteredverticalbar() {
 
     chart._getName = function () {
         return _NAME;
+    }
+
+    chart._getHTML = function () {
+        return _local_svg.node().outerHTML;
     }
 
     chart.update = function (data) {
@@ -849,19 +914,40 @@ function clusteredverticalbar() {
         var xAxisGroup,
             yAxisGroup;
 
-        xAxisGroup = plot.select('.x.axis')
-            .transition()
-            .duration(COMMON.DURATION)
-            .call(d3.axisBottom(x0));
+        var isRotate = false;
 
-        _setAxisColor(xAxisGroup, _xAxisColor);
+        if (_showXaxis) {
+            xAxisGroup = plot.select('.x.axis')
+                .transition()
+                .duration(COMMON.DURATION)
+                .call(d3.axisBottom(x0)
+                    .tickFormat(function (d) {
+                        if (isRotate == false) {
+                            isRotate = UTIL.getTickRotate(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                        }
+                        return UTIL.getTruncatedTick(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                    }));
 
-        yAxisGroup = plot.select('.y.axis')
-            .transition()
-            .duration(COMMON.DURATION)
-            .call(d3.axisLeft(y).ticks(null, "s"));
+            _setAxisColor(xAxisGroup, _xAxisColor);
 
-        _setAxisColor(yAxisGroup, _yAxisColor);
+            if (isRotate) {
+                _local_svg.selectAll('.x .tick text')
+                    .attr("transform", "rotate(-15)");
+            }
+            else {
+                _local_svg.selectAll('.x .tick text')
+                    .attr("transform", "rotate(0)");
+            }
+        }
+
+        if (_showYaxis) {
+            yAxisGroup = plot.select('.y.axis')
+                .transition()
+                .duration(COMMON.DURATION)
+                .call(d3.axisLeft(y).ticks(null, "s"));
+
+            _setAxisColor(yAxisGroup, _yAxisColor);
+        }
 
         /* Update Axes Grid */
         _localXGrid.ticks(_localXLabels.length);
@@ -966,6 +1052,14 @@ function clusteredverticalbar() {
         return chart;
     }
 
+    chart.showGrid = function (value) {
+        if (!arguments.length) {
+            return _showGrid;
+        }
+        _showGrid = value;
+        return chart;
+    }
+
     chart.showYaxisLabel = function (value) {
         if (!arguments.length) {
             return _showYaxisLabel;
@@ -992,7 +1086,7 @@ function clusteredverticalbar() {
 
     chart.displayName = function (value) {
         if (!arguments.length) {
-            return _tooltip;
+            return _displayName;
         }
         _displayName = value;
         return chart;
@@ -1004,6 +1098,14 @@ function clusteredverticalbar() {
             measureName: measureName
         }
         return _legendData;
+    }
+
+    chart.print = function (value) {
+        if (!arguments.length) {
+            return _print;
+        }
+        _print = value;
+        return chart;
     }
 
     chart.showValues = function (value, measure) {
