+++ js/charts/heatmap.js
@@ -297,7 +297,7 @@ function heatmap() {
         }
     }
 
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
@@ -305,9 +305,11 @@ function heatmap() {
         }
     }
 
-    var clearFilter = function () {
+   var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
         }
     }
 
@@ -518,10 +520,10 @@ function heatmap() {
             drawViz(cell);
 
             d3.select(div).select('.filterData')
-                .on('click', applyFilter(chart));
+                .on('click', applyFilter(div));
 
             d3.select(div).select('.removeFilter')
-                .on('click', clearFilter());
+                .on('click', clearFilter(div));
 
             var lasso = d3.lasso()
                 .hoverSelect(true)
