+++ js/charts/line.js
@@ -1,6 +1,8 @@
-var COMMON = require('../extras/common.js')(),
-    UTIL = require('../extras/util.js')(),
-    LEGEND = require('../extras/legend_barcharts.js')();
+//var d3 = require('d3');
+var COMMON = require('../extras/common.js')();
+var UTIL = require('../extras/util.js')();
+var LEGEND = require('../extras/legend_barcharts.js')();
+//var LASSO = require('d3-lasso'); // creates a new lasso
 
 function line() {
 
@@ -17,11 +19,11 @@ function line() {
         _showYaxis,
         _showXaxisLabel,
         _showYaxisLabel,
+        _showGrid,
         _xAxisColor,
         _yAxisColor,
         _displayName,
         _legendData,
-
         _showValues = [],
         _displayNameForMeasure = [],
         _fontStyle = [],
@@ -32,7 +34,8 @@ function line() {
         _borderColor = [],
         _fontSize = [],
         _lineType = [],
-        _pointType = [];
+        _pointType = [],
+        _print;
 
     var margin = {
         top: 0,
@@ -42,6 +45,15 @@ function line() {
     };
 
     var _local_svg, _Local_data, _originalData, _localLabelStack = [], legendBreakCount = 1;
+    var _localXAxis,
+        _localYAxis,
+        _localXGrid,
+        _localYGrid;
+
+    var tickLength = d3.scaleLinear()
+        .domain([22, 34])
+        .range([4, 6]);
+
     var legendSpace = 20, axisLabelSpace = 20, offsetX = 16, offsetY = 3, div;
     var parentWidth, parentHeight, plotWidth, plotHeight, container;
     var threshold = [];
@@ -52,18 +64,15 @@ function line() {
         this.measure(config.measure);
         this.showLegend(config.showLegend);
         this.legendPosition(config.legendPosition);
-
         this.showXaxis(config.showXaxis);
         this.showYaxis(config.showYaxis);
         this.showXaxisLabel(config.showXaxisLabel);
-
         this.showYaxisLabel(config.showYaxisLabel);
         this.xAxisColor(config.xAxisColor);
         this.yAxisColor(config.yAxisColor);
         this.displayName(config.displayName);
-
+        this.showGrid(config.showGrid);
         this.showXaxisLabel(config.showXaxisLabel);
-
         this.showValues(config.showValues);
         this.displayNameForMeasure(config.displayNameForMeasure);
         this.fontStyle(config.fontStyle);
@@ -75,7 +84,6 @@ function line() {
         this.fontSize(config.fontSize);
         this.lineType(config.lineType);
         this.pointType(config.pointType);
-
         this.legendData(config.displayColor, config.measure);
     }
     var getPointType = function (index) {
@@ -150,6 +158,7 @@ function line() {
             }
         }
     }
+
     var onLassoDraw = function (lasso, scope) {
         return function () {
             filter = true;
@@ -165,6 +174,7 @@ function line() {
                 .classed('possible', false);
         }
     }
+
     var onLassoEnd = function (lasso, scope) {
         return function () {
             var data = lasso.selectedItems().data();
@@ -200,16 +210,18 @@ function line() {
             }
         }
     }
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
             }
         }
     }
-    var clearFilter = function () {
+   var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
         }
     }
     var _handleMouseOverFn = function (tooltip, container) {
@@ -249,6 +261,20 @@ function line() {
             }
         }
     }
+
+    var _setAxisColor = function (axis, color) {
+        var path = axis.select('path'),
+            ticks = axis.selectAll('.tick');
+
+        path.style('stroke', color);
+
+        ticks.select('line')
+            .style('stroke', color);
+
+        ticks.select('text')
+            .style('fill', color);
+    }
+
     function chart(selection) {
         _local_svg = selection;
 
@@ -411,6 +437,38 @@ function line() {
             });
         })]).nice();
 
+        var _localXLabels = data.map(function (d) {
+            return d[_dimension[0]];
+        });
+
+        _localXGrid = d3.axisBottom()
+            .ticks(_localXLabels.length)
+            .tickFormat('')
+            .tickSize(-plotHeight);
+
+        _localYGrid = d3.axisLeft()
+            .tickFormat('')
+            .tickSize(-plotWidth);
+
+        _localXGrid.scale(x);
+        _localYGrid.scale(y);
+
+        plot.append('g')
+            .attr('class', 'x grid')
+            .attr('visibility', function () {
+                return _showGrid ? 'visible' : 'hidden';
+            })
+            .attr('transform', 'translate(0, ' + plotHeight + ')')
+            .call(_localXGrid);
+
+        plot.append('g')
+            .attr('class', 'y grid')
+            .attr('visibility', function () {
+                return _showGrid ? 'visible' : 'hidden';
+            })
+            .call(_localYGrid);
+
+
         var areaGenerator = d3.area()
             .curve(d3.curveLinear)
             .x(function (d, i) {
@@ -441,7 +499,7 @@ function line() {
             })
             .attr('class', 'area')
             .attr('fill', function (d, i) {
-                return UTIL.getDisplayColor(_measure.indexOf(d[0]['tag']), _displayColor);
+                return UTIL.getBorderColor(_measure.indexOf(d[0]['tag']), _borderColor);
             })
             .attr('visibility', function (d, i) {
                 if (_lineType[(_measure.indexOf(d[0]['tag']))] == "Area") {
@@ -450,45 +508,50 @@ function line() {
                 else {
                     return 'hidden';
                 }
-
-            })
-            .on("mouseover", function (d) {
-                d3.select(this)
-                    .style("fill-opacity", 1)
-                    .style("cursor", "pointer");
-            })
-            .on("mouseout", function (d) {
-                d3.select(this)
-                    .style("fill-opacity", .5)
-                    .style("cursor", "none");
             })
             .style('fill-opacity', 0.5)
             .attr('stroke', 'none')
             .attr('d', areaGenerator);
 
         var line = clusterLine.append('path')
+            .classed('line-path', true)
             .datum(function (d, i) {
                 return data.map(function (datum) { return { "tag": d, "data": datum }; });
             })
             .attr('class', 'line')
-            .attr('fill', 'none')
+            .style('fill', 'none')
             .attr('stroke', function (d, i) {
                 return UTIL.getDisplayColor(_measure.indexOf(d[0]['tag']), _displayColor);
             })
             .attr('stroke-linejoin', 'round')
             .attr('stroke-linecap', 'round')
             .attr('stroke-width', 1)
-            .on("mouseover", function (d) {
-                d3.select(this)
-                    .style("stroke-width", "2.5px")
-                    .style("cursor", "pointer");
-            })
-            .on("mouseout", function (d) {
-                d3.select(this)
-                    .style("stroke-width", "1.5px")
-                    .style("cursor", "none");
-            })
-            .attr('d', lineGenerator);
+
+        if (!_print) {
+            line
+                .on("mouseover", function (d) {
+                    d3.select(this)
+                        .style("stroke-width", "2.5px")
+                        .style("cursor", "pointer");
+                })
+                .on("mouseout", function (d) {
+                    d3.select(this)
+                        .style("stroke-width", "1.5px")
+                        .style("cursor", "none");
+                })
+                .attr('d', lineGenerator)
+                .transition()
+                .duration(COMMON.DURATION)
+                .attrTween('stroke-dasharray', function () {
+                    var l = this.getTotalLength(),
+                        interpolator = d3.interpolateString('0,' + l, l + ',' + l);
+
+                    return function (t) {
+                        return interpolator(t);
+                    }
+                });
+        }
+
 
         var point = clusterLine.selectAll('point')
             .data(function (d, i) {
@@ -500,7 +563,7 @@ function line() {
                 return UTIL.getDisplayColor(_measure.indexOf(d.tag), _displayColor);
             })
             .attr('fill', function (d, i) {
-                return  UTIL.getBorderColor(_measure.indexOf(d.tag), _borderColor);
+                return UTIL.getBorderColor(_measure.indexOf(d.tag), _borderColor);
             })
             .attr('d', function (d, i) {
                 return d3.symbol()
@@ -512,76 +575,122 @@ function line() {
                     + (x(d['data'][_dimension[0]]) + x.bandwidth() / 2)
                     + ',' + y(d['data'][d['tag']]) + ')';
             })
-            .on('mouseover', _handleMouseOverFn.call(chart, tooltip, _local_svg))
-            .on('mousemove', _handleMouseMoveFn.call(chart, tooltip, _local_svg))
-            .on('mouseout', _handleMouseOutFn.call(chart, tooltip, _local_svg))
-            .on('click', function (d) {
-                if ($("#myonoffswitch").prop('checked') == false) {
-                    $('#Modal_' + $(div).attr('id') + ' .measure').val(d.measure);
-                    $('#Modal_' + $(div).attr('id') + ' .threshold').val('');
-                    $('#Modal_' + $(div).attr('id') + ' .measure').attr('disabled', true);;
-                    $('#Modal_' + $(div).attr('id')).modal('toggle');
-                }
-                else {
-                    var confirm = d3.select(div).select('.confirm')
-                        .style('visibility', 'visible');
-                    var _filter = _Local_data.filter(function (d1) {
-                        return d.data[_dimension[0]] === d1[_dimension[0]]
-                    })
-                    var rect = d3.select(this);
-                    if (rect.classed('selected')) {
-                        rect.classed('selected', false);
-                        filterData.map(function (val, i) {
-                            if (val[_dimension[0]] == d.data[_dimension[0]]) {
-                                filterData.splice(i, 1)
-                            }
+
+        if (!_print) {
+            point.on('mouseover', _handleMouseOverFn.call(chart, tooltip, _local_svg))
+                .on('mousemove', _handleMouseMoveFn.call(chart, tooltip, _local_svg))
+                .on('mouseout', _handleMouseOutFn.call(chart, tooltip, _local_svg))
+                .on('click', function (d) {
+                    if ($("#myonoffswitch").prop('checked') == false) {
+                        $('#Modal_' + $(div).attr('id') + ' .measure').val(d.measure);
+                        $('#Modal_' + $(div).attr('id') + ' .threshold').val('');
+                        $('#Modal_' + $(div).attr('id') + ' .measure').attr('disabled', true);;
+                        $('#Modal_' + $(div).attr('id')).modal('toggle');
+                    }
+                    else {
+                        var confirm = d3.select(div).select('.confirm')
+                            .style('visibility', 'visible');
+                        var _filter = _Local_data.filter(function (d1) {
+                            return d.data[_dimension[0]] === d1[_dimension[0]]
                         })
-                    } else {
-                        rect.classed('selected', true);
-                        var isExist = filterData.filter(function (val) {
-                            if (val[_dimension[0]] == d.data[_dimension[0]]) {
-                                return val
+                        var rect = d3.select(this);
+                        if (rect.classed('selected')) {
+                            rect.classed('selected', false);
+                            filterData.map(function (val, i) {
+                                if (val[_dimension[0]] == d.data[_dimension[0]]) {
+                                    filterData.splice(i, 1)
+                                }
+                            })
+                        } else {
+                            rect.classed('selected', true);
+                            var isExist = filterData.filter(function (val) {
+                                if (val[_dimension[0]] == d.data[_dimension[0]]) {
+                                    return val
+                                }
+                            })
+                            if (isExist.length == 0) {
+                                filterData.push(_filter[0]);
                             }
-                        })
-                        if (isExist.length == 0) {
-                            filterData.push(_filter[0]);
                         }
                     }
-                }
-            })
+                })
+        }
 
-        plot.append("g")
-            .attr("class", "x_axis")
-            .attr("transform", "translate(0," + plotHeight + ")")
-            .call(d3.axisBottom(x))
-            .append("text")
-            .attr("x", plotWidth / 2)
-            .attr("y", 2 * axisLabelSpace)
-            .attr("dy", "0.32em")
-            .attr("fill", "#000")
-            .attr("font-weight", "bold")
-            .style('text-anchor', 'middle')
-            .style('visibility', UTIL.getVisibility(_showXaxisLabel))
-            .text(function () {
-                return _displayName;
-            });
+        /* Axes */
+        var xAxisGroup,
+            yAxisGroup;
 
-        plot.append("g")
-            .attr("class", "y_axis")
-            .call(d3.axisLeft(y).ticks(null, "s"))
-            .append("text")
-            .attr("x", plotHeight / 2)
-            .attr("y", 2 * axisLabelSpace)
-            .attr("transform", function (d) { return "rotate(" + 90 + ")"; })
-            .attr("dy", "0.32em")
-            .style('visibility', UTIL.getVisibility(_showYaxisLabel))
-            .attr("font-weight", "bold")
-            .style('text-anchor', 'middle')
-            .text(function () {
-                return _displayNameForMeasure.map(function (p) { return p; }).join(', ');
-            });
+        var isRotate = false;
 
-        UTIL.setAxisColor(_local_svg, _yAxisColor, _xAxisColor, _showYaxis, _showXaxis, _showYaxis, _showXaxis);
+        if (_showXaxis) {
+            _localXAxis = d3.axisBottom(x)
+                .tickSize(0)
+                .tickFormat(function (d) {
+                    if (isRotate == false) {
+                        isRotate = UTIL.getTickRotate(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                    }
+                    return UTIL.getTruncatedTick(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                })
+                .tickPadding(10);
+
+            xAxisGroup = plot.append('g')
+                .attr('class', 'x axis')
+                .attr('visibility', function () {
+                    return 'visible';
+                })
+                .attr('transform', 'translate(0, ' + plotHeight + ')')
+                .call(_localXAxis);
+
+            xAxisGroup.append('g')
+                .attr('class', 'label')
+                .attr('transform', function () {
+                    return 'translate(' + (plotWidth) + ', ' + (COMMON.AXIS_THICKNESS / 1.5) + ')';
+                })
+                .append('text')
+                .style('text-anchor', 'end')
+                .style('font-weight', 'bold')
+                .style('fill', _xAxisColor)
+                .text(_displayName);
+
+            if (isRotate) {
+                _local_svg.selectAll('.x .tick text')
+                    .attr("transform", "rotate(-15)");
+            }
+            _setAxisColor(xAxisGroup, _xAxisColor);
+        }
+
+        if (_showYaxis) {
+            _localYAxis = d3.axisLeft(y)
+                .tickSize(0)
+                .tickPadding(8)
+                .tickFormat(function (d) {
+                    return UTIL.shortScale(2)(d);
+                });
+
+            yAxisGroup = plot.append('g')
+                .attr('class', 'y axis')
+                .attr('visibility', function () {
+                    return 'visible';
+                })
+                .call(_localYAxis);
+
+            yAxisGroup.append('g')
+                .attr('class', 'label')
+                .attr('transform', function () {
+                    return 'translate(' + (-COMMON.AXIS_THICKNESS / 1.15) + ', ' + '0)';
+                })
+                .append('text')
+                .attr('transform', 'rotate(-90)')
+                .style('text-anchor', 'end')
+                .style('font-weight', 'bold')
+                .style('fill', _yAxisColor)
+                .text(function () {
+                    return _displayNameForMeasure.map(function (p) { return p; }).join(', ');
+                });
+
+            _setAxisColor(yAxisGroup, _yAxisColor);
+        }
+
 
         _local_svg.select('g.sort').remove();
         UTIL.sortingView(container, parentHeight, parentWidth + margin.left, legendBreakCount, axisLabelSpace, offsetX);
@@ -606,10 +715,10 @@ function line() {
             });
 
         d3.select(div).select('.filterData')
-            .on('click', applyFilter(chart));
+            .on('click', applyFilter(div));
 
         d3.select(div).select('.removeFilter')
-            .on('click', clearFilter());
+            .on('click', clearFilter(div));
 
         var lasso = d3.lasso()
             .hoverSelect(true)
@@ -625,7 +734,11 @@ function line() {
         _local_svg.call(lasso);
     }
 
- chart._legendInteraction = function (event, data, plot) {
+    chart._legendInteraction = function (event, data, plot) {
+        if (_print) {
+            // No interaction during print enabled
+            return;
+        }
         switch (event) {
             case 'mouseover':
                 _legendMouseOver(data, plot);
@@ -641,7 +754,7 @@ function line() {
                 break;
         }
     }
-     var _legendMouseOver = function (data, plot) {
+    var _legendMouseOver = function (data, plot) {
 
         var line = plot.selectAll('.line')
             .filter(function (d, i) {
@@ -651,18 +764,18 @@ function line() {
             .style('stroke', COMMON.HIGHLIGHTER);
     }
 
-     var _legendMouseMove = function (data, plot) {
+    var _legendMouseMove = function (data, plot) {
 
     }
 
-     var _legendMouseOut = function (data, plot) {
+    var _legendMouseOut = function (data, plot) {
         var line = plot.selectAll('.line')
             .filter(function (d, i) {
                 return d[i].tag === data;
             })
             .style("stroke-width", "1.5px")
             .style('stroke', function (d, i) {
-                return UTIL.getBorderColor(_measure.indexOf(d[0]['tag']), _borderColor);
+                return UTIL.getDisplayColor(_measure.indexOf(d[0]['tag']), _displayColor);
             });
     }
 
@@ -686,6 +799,10 @@ function line() {
 
         var y = d3.scaleLinear()
             .rangeRound([plotHeight, 0]);
+
+        var _localXLabels = data.map(function (d) {
+            return d[_dimension[0]];
+        });
 
         var keys = UTIL.getMeasureList(data[0], _dimension);
 
@@ -733,15 +850,63 @@ function line() {
         plot.selectAll('path.point').remove()
 
 
-        plot.select('.x_axis')
-            .transition()
-            .duration(COMMON.DURATION)
-            .call(d3.axisBottom(x));
+        var xAxisGroup,
+            yAxisGroup;
 
-        plot.select('.y_axis')
+        var isRotate = false;
+
+        if (_showXaxis) {
+            xAxisGroup = plot.select('.x.axis')
+                .transition()
+                .duration(COMMON.DURATION)
+                .call(d3.axisBottom(x)
+                    .tickFormat(function (d) {
+                        if (isRotate == false) {
+                            isRotate = UTIL.getTickRotate(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                        }
+                        return UTIL.getTruncatedTick(d, (plotWidth) / (_localXLabels.length - 1), tickLength);
+                    }));
+
+            _setAxisColor(xAxisGroup, _xAxisColor);
+
+            if (isRotate) {
+                _local_svg.selectAll('.x .tick text')
+                    .attr("transform", "rotate(-15)");
+            }
+            else {
+                _local_svg.selectAll('.x .tick text')
+                    .attr("transform", "rotate(0)");
+            }
+        }
+
+        if (_showYaxis) {
+            yAxisGroup = plot.select('.y.axis')
+                .transition()
+                .duration(COMMON.DURATION)
+                .call(d3.axisLeft(y).ticks(null, "s"));
+
+            _setAxisColor(yAxisGroup, _yAxisColor);
+        }
+
+        /* Update Axes Grid */
+        _localXGrid.ticks(x);
+        _localYGrid.scale(y);
+
+        plot.select('.x.grid')
             .transition()
             .duration(COMMON.DURATION)
-            .call(d3.axisLeft(y).ticks(null, "s"));
+            .attr('visibility', function () {
+                return _showGrid ? 'visible' : 'hidden';
+            })
+            .call(_localXGrid);
+
+        plot.select('.y.grid')
+            .transition()
+            .duration(COMMON.DURATION)
+            .attr('visibility', function () {
+                return _showGrid ? 'visible' : 'hidden';
+            })
+            .call(_localYGrid);
 
         UTIL.setAxisColor(_local_svg, _yAxisColor, _xAxisColor, _showYaxis, _showXaxis);
         UTIL.displayThreshold(threshold, data, keys);
@@ -750,6 +915,10 @@ function line() {
 
     chart._getName = function () {
         return _NAME;
+    }
+
+    chart._getHTML = function () {
+        return _local_svg.node().outerHTML;
     }
 
     chart.config = function (value) {
@@ -859,7 +1028,7 @@ function line() {
 
     chart.showGrid = function (value) {
         if (!arguments.length) {
-            return _tooltip;
+            return _showGrid;
         }
         _showGrid = value;
         return chart;
@@ -867,7 +1036,7 @@ function line() {
 
     chart.stacked = function (value) {
         if (!arguments.length) {
-            return _tooltip;
+            return _stacked;
         }
         _stacked = value;
         return chart;
@@ -875,9 +1044,17 @@ function line() {
 
     chart.displayName = function (value) {
         if (!arguments.length) {
-            return _tooltip;
+            return _displayName;
         }
         _displayName = value;
+        return chart;
+    }
+
+    chart.showGrid = function (value) {
+        if (!arguments.length) {
+            return _showGrid;
+        }
+        _showGrid = value;
         return chart;
     }
 
@@ -887,6 +1064,14 @@ function line() {
             measureName: measureName
         }
         return _legendData;
+    }
+
+    chart.print = function (value) {
+        if (!arguments.length) {
+            return _print;
+        }
+        _print = value;
+        return chart;
     }
 
     chart.showValues = function (value, measure) {
