+++ js/charts/sankey.js
@@ -168,7 +168,7 @@ function sankey() {
         }
     }
 
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
@@ -176,9 +176,11 @@ function sankey() {
         }
     }
 
-    var clearFilter = function () {
+   var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
         }
     }
 
@@ -500,10 +502,10 @@ function sankey() {
                 .attr('text-anchor', 'start');
 
             d3.select(div).select('.filterData')
-                .on('click', applyFilter(chart));
+                .on('click', applyFilter(div));
 
             d3.select(div).select('.removeFilter')
-                .on('click', clearFilter());
+                .on('click', clearFilter(div));
 
             var lasso = d3.lasso()
                 .hoverSelect(true)
