+++ js/charts/treemap.js
@@ -419,7 +419,7 @@ function treemap() {
             }
         }
     }
-    var applyFilter = function (chart) {
+    var applyFilter = function (div) {
         return function () {
             if (filterData.length > 0) {
                 chart.update(filterData);
@@ -427,9 +427,11 @@ function treemap() {
         }
     }
 
-    var clearFilter = function () {
+   var clearFilter = function (div) {
         return function () {
             chart.update(_originalData);
+            d3.select(div).select('.confirm')
+                .style('visibility', 'hidden');
         }
     }
 
@@ -649,10 +651,10 @@ function treemap() {
             drawViz(cell)
 
             d3.select(div).select('.filterData')
-                .on('click', applyFilter(chart));
+                .on('click', applyFilter(div));
 
             d3.select(div).select('.removeFilter')
-                .on('click', clearFilter());
+                .on('click', clearFilter(div));
 
             var lasso = d3.lasso()
                 .hoverSelect(true)
